refactor: switch to domain driven file structure
