+++ apps/backend/api/support-resource/subschemas/availability-pattern.schema.ts
@@ -1,7 +1,7 @@
 import { Schema } from "mongoose";
 import { IAvailabilityPattern } from "@shared/database/SupportResource.types";
-import { TimeSchema } from "./Time";
-import { RRuleSchema } from "./RRule";
+import { TimeSchema } from "./time.schema";
+import { RRuleSchema } from "./rrule.schema";
 
 export const AvailabilityPatternSchema: Schema<IAvailabilityPattern> =
   new Schema(
