--- apps/backend/api/types/Agreement.types.ts
@@ -1,10 +1,0 @@
-export interface IAgreement {
-  id: string;
-  code: string;
-  origin_code: string;
-  service_type: string;
-  service_name: string;
-  amount: number;
-  provider_code: string;
-  year: number;
-}
